requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
datasets>=2.14.0
trl>=0.7.0
llama-cpp-python>=0.2.0
psutil>=5.9.0
numpy>=1.24.0
accelerate>=0.25.0
scipy>=1.11.0
requests>=2.31.0
sentencepiece>=0.1.99
protobuf>=4.25.0
